BigInteger has been changed into BigDecimal.

+++ src/com/company/Main.java
@@ -1,22 +1,22 @@
 package com.company;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.InputMismatchException;
 import java.util.Scanner;
 
 public class Main {
-    public static void collatzConjecture(BigInteger number, int odd, int even) {
-        if ((number.remainder(BigInteger.valueOf(2))).equals(BigInteger.valueOf(0))) {
+    public static void collatzConjecture(BigDecimal number, int odd, int even) {
+        if ((number.remainder(BigDecimal.valueOf(2))).equals(BigDecimal.valueOf(0))) {
             ++even;
-            number = number.divide(BigInteger.valueOf(2));
+            number = number.divide(BigDecimal.valueOf(2));
             System.out.println(number);
             collatzConjecture(number, odd, even);
-        } else if (number.equals(BigInteger.valueOf(1))) {
+        } else if (number.equals(BigDecimal.valueOf(1))) {
             System.out.print("You are into a loop >4->2->1-\n");
             System.out.println("Odd: " + odd + "; Even: " + even);
         } else {
             ++odd;
-            number = number.multiply(BigInteger.valueOf(3)).add(BigInteger.valueOf(1));
+            number = number.multiply(BigDecimal.valueOf(3)).add(BigDecimal.valueOf(1));
             System.out.println(number);
             collatzConjecture(number, odd, even);
         }
@@ -30,7 +30,7 @@ public class Main {
         while (true) {
             try {
                 System.out.print("Type in number from 1+ : ");
-                BigInteger number = scanner.nextBigInteger();
+                BigDecimal number = scanner.nextBigDecimal();
                 if (number.intValue() <= 0) {
                     throw new InputMismatchException();
                 }
